Put the playground description in a separate source file
