--- Readme.playground/timeline.xctimeline
@@ -1,46 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<Timeline
-   version = "3.0">
-   <TimelineItems>
-      <LoggerValueHistoryTimelineItem
-         documentLocation = "file:///Users/marijnschilling/Projects/TinyConstraints/Readme.playground#CharacterRangeLen=115&amp;CharacterRangeLoc=1240&amp;EndingColumnNumber=7&amp;EndingLineNumber=49&amp;StartingColumnNumber=27&amp;StartingLineNumber=43&amp;Timestamp=551387982.3195879"
-         selectedRepresentationIndex = "0"
-         shouldTrackSuperviewWidth = "NO">
-      </LoggerValueHistoryTimelineItem>
-      <LoggerValueHistoryTimelineItem
-         documentLocation = "file:///Users/marijnschilling/Projects/TinyConstraints/Readme.playground#CharacterRangeLen=1&amp;CharacterRangeLoc=838&amp;EndingColumnNumber=12&amp;EndingLineNumber=24&amp;StartingColumnNumber=11&amp;StartingLineNumber=24&amp;Timestamp=551387982.319764"
-         selectedRepresentationIndex = "0"
-         shouldTrackSuperviewWidth = "NO">
-      </LoggerValueHistoryTimelineItem>
-      <LoggerValueHistoryTimelineItem
-         documentLocation = "file:///Users/marijnschilling/Projects/TinyConstraints/Readme.playground#CharacterRangeLen=1&amp;CharacterRangeLoc=838&amp;EndingColumnNumber=12&amp;EndingLineNumber=24&amp;StartingColumnNumber=11&amp;StartingLineNumber=24&amp;Timestamp=551387982.319875"
-         selectedRepresentationIndex = "0"
-         shouldTrackSuperviewWidth = "NO">
-      </LoggerValueHistoryTimelineItem>
-      <LoggerValueHistoryTimelineItem
-         documentLocation = "file:///Users/marijnschilling/Projects/TinyConstraints/Readme.playground#CharacterRangeLen=51&amp;CharacterRangeLoc=761&amp;EndingColumnNumber=12&amp;EndingLineNumber=22&amp;StartingColumnNumber=43&amp;StartingLineNumber=20&amp;Timestamp=551387982.319981"
-         selectedRepresentationIndex = "0"
-         shouldTrackSuperviewWidth = "NO">
-      </LoggerValueHistoryTimelineItem>
-      <LoggerValueHistoryTimelineItem
-         documentLocation = "file:///Users/marijnschilling/Projects/TinyConstraints/Readme.playground#CharacterRangeLen=89&amp;CharacterRangeLoc=1136&amp;EndingColumnNumber=12&amp;EndingLineNumber=43&amp;StartingColumnNumber=12&amp;StartingLineNumber=37&amp;Timestamp=551387982.3200819"
-         selectedRepresentationIndex = "0"
-         shouldTrackSuperviewWidth = "NO">
-      </LoggerValueHistoryTimelineItem>
-      <LoggerValueHistoryTimelineItem
-         documentLocation = "file:///Users/marijnschilling/Projects/TinyConstraints/Readme.playground#CharacterRangeLen=1&amp;CharacterRangeLoc=838&amp;EndingColumnNumber=12&amp;EndingLineNumber=24&amp;StartingColumnNumber=11&amp;StartingLineNumber=24&amp;Timestamp=551387982.320181"
-         selectedRepresentationIndex = "0"
-         shouldTrackSuperviewWidth = "NO">
-      </LoggerValueHistoryTimelineItem>
-      <LoggerValueHistoryTimelineItem
-         documentLocation = "file:///Users/marijnschilling/Projects/TinyConstraints/Readme.playground#CharacterRangeLen=2&amp;CharacterRangeLoc=1937&amp;EndingColumnNumber=17&amp;EndingLineNumber=66&amp;StartingColumnNumber=15&amp;StartingLineNumber=66&amp;Timestamp=551387982.32028"
-         selectedRepresentationIndex = "0"
-         shouldTrackSuperviewWidth = "NO">
-      </LoggerValueHistoryTimelineItem>
-      <LoggerValueHistoryTimelineItem
-         documentLocation = "file:///Users/marijnschilling/Projects/TinyConstraints/Readme.playground#CharacterRangeLen=10&amp;CharacterRangeLoc=1990&amp;EndingColumnNumber=2&amp;EndingLineNumber=69&amp;StartingColumnNumber=31&amp;StartingLineNumber=68&amp;Timestamp=551387982.320376"
-         selectedRepresentationIndex = "0"
-         shouldTrackSuperviewWidth = "NO">
-      </LoggerValueHistoryTimelineItem>
-   </TimelineItems>
-</Timeline>
