+++ TinyConstraints.playground/Pages/0 - Tiny Constraints.xcplaygroundpage/Contents.swift
@@ -37,39 +37,3 @@ PlaygroundPage.current.liveView = viewController
 
 //: with TinyConstraints:
 
-
-
-
-extension Constraint: CustomPlaygroundDisplayConvertible {
-    public var playgroundDescription: Any {
-
-        guard secondItem != nil else {
-            return onlyOneAttributeView
-        }
-
-        let superView = UIView(frame: CGRect(x: 0, y: 0, width: 200, height: 200))
-        superView.backgroundColor = .red
-
-
-        return "lalala"
-    }
-
-    var onlyOneAttributeView: Any {
-        let view = UIView(frame: CGRect(x: 0, y: 0, width: 203, height: 203))
-        let imageView = UIImageView(image: UIImage(named: "height&width.png")!)
-
-        let label = UILabel(frame: CGRect(x: 0, y: 90, width: 200, height: 20))
-        label.textColor = .white
-        label.textAlignment = .center
-        label.text = "\(constant)"
-
-        view.addSubview(imageView)
-        imageView.addSubview(label)
-
-        if firstAttribute == .height {
-            imageView.transform = CGAffineTransform(rotationAngle: CGFloat(Double.pi * 1.5))
-        }
-
-        return view
-    }
-}
