+++ TinyConstraints.playground/Pages/0 - Tiny Constraints.xcplaygroundpage/timeline.xctimeline
@@ -1,0 +1,51 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<Timeline
+   version = "3.0">
+   <TimelineItems>
+      <LoggerValueHistoryTimelineItem
+         documentLocation = "file:///Users/marijnschilling/Projects/TinyConstraints/TinyConstraints.playground/Pages/0%20-%20Tiny%20Constraints.xcplaygroundpage#CharacterRangeLen=0&amp;CharacterRangeLoc=1151&amp;EndingColumnNumber=0&amp;EndingLineNumber=39&amp;StartingColumnNumber=0&amp;StartingLineNumber=39&amp;Timestamp=552134501.133431"
+         selectedRepresentationIndex = "0"
+         shouldTrackSuperviewWidth = "NO">
+      </LoggerValueHistoryTimelineItem>
+      <LoggerValueHistoryTimelineItem
+         documentLocation = "file:///Users/marijnschilling/Projects/TinyConstraints/TinyConstraints.playground/Pages/0%20-%20Tiny%20Constraints.xcplaygroundpage#CharacterRangeLen=1&amp;CharacterRangeLoc=838&amp;EndingColumnNumber=12&amp;EndingLineNumber=24&amp;StartingColumnNumber=11&amp;StartingLineNumber=24&amp;Timestamp=552134501.133944"
+         selectedRepresentationIndex = "0"
+         shouldTrackSuperviewWidth = "NO">
+      </LoggerValueHistoryTimelineItem>
+      <LoggerValueHistoryTimelineItem
+         documentLocation = "file:///Users/marijnschilling/Projects/TinyConstraints/TinyConstraints.playground/Pages/0%20-%20Tiny%20Constraints.xcplaygroundpage#CharacterRangeLen=1&amp;CharacterRangeLoc=838&amp;EndingColumnNumber=12&amp;EndingLineNumber=24&amp;StartingColumnNumber=11&amp;StartingLineNumber=24&amp;Timestamp=552134501.134164"
+         selectedRepresentationIndex = "0"
+         shouldTrackSuperviewWidth = "NO">
+      </LoggerValueHistoryTimelineItem>
+      <LoggerValueHistoryTimelineItem
+         documentLocation = "file:///Users/marijnschilling/Projects/TinyConstraints/TinyConstraints.playground/Pages/0%20-%20Tiny%20Constraints.xcplaygroundpage#CharacterRangeLen=51&amp;CharacterRangeLoc=761&amp;EndingColumnNumber=12&amp;EndingLineNumber=22&amp;StartingColumnNumber=43&amp;StartingLineNumber=20&amp;Timestamp=552134501.134365"
+         selectedRepresentationIndex = "0"
+         shouldTrackSuperviewWidth = "NO">
+      </LoggerValueHistoryTimelineItem>
+      <LoggerValueHistoryTimelineItem
+         documentLocation = "file:///Users/marijnschilling/Projects/TinyConstraints/TinyConstraints.playground/Pages/0%20-%20Tiny%20Constraints.xcplaygroundpage#CharacterRangeLen=15&amp;CharacterRangeLoc=1136&amp;EndingColumnNumber=0&amp;EndingLineNumber=39&amp;StartingColumnNumber=12&amp;StartingLineNumber=37&amp;Timestamp=552134501.13455"
+         selectedRepresentationIndex = "0"
+         shouldTrackSuperviewWidth = "NO">
+      </LoggerValueHistoryTimelineItem>
+      <LoggerValueHistoryTimelineItem
+         documentLocation = "file:///Users/marijnschilling/Projects/TinyConstraints/TinyConstraints.playground/Pages/0%20-%20Tiny%20Constraints.xcplaygroundpage#CharacterRangeLen=1&amp;CharacterRangeLoc=838&amp;EndingColumnNumber=12&amp;EndingLineNumber=24&amp;StartingColumnNumber=11&amp;StartingLineNumber=24&amp;Timestamp=552134501.1347359"
+         selectedRepresentationIndex = "0"
+         shouldTrackSuperviewWidth = "NO">
+      </LoggerValueHistoryTimelineItem>
+      <LoggerValueHistoryTimelineItem
+         documentLocation = "file:///Users/marijnschilling/Projects/TinyConstraints/TinyConstraints.playground/Pages/0%20-%20Tiny%20Constraints.xcplaygroundpage#CharacterRangeLen=0&amp;CharacterRangeLoc=1151&amp;EndingColumnNumber=0&amp;EndingLineNumber=39&amp;StartingColumnNumber=0&amp;StartingLineNumber=39&amp;Timestamp=552134501.134908"
+         selectedRepresentationIndex = "0"
+         shouldTrackSuperviewWidth = "NO">
+      </LoggerValueHistoryTimelineItem>
+      <LoggerValueHistoryTimelineItem
+         documentLocation = "file:///Users/marijnschilling/Projects/TinyConstraints/TinyConstraints.playground/Pages/0%20-%20Tiny%20Constraints.xcplaygroundpage#CharacterRangeLen=0&amp;CharacterRangeLoc=1151&amp;EndingColumnNumber=0&amp;EndingLineNumber=39&amp;StartingColumnNumber=0&amp;StartingLineNumber=39&amp;Timestamp=552134501.1350909"
+         selectedRepresentationIndex = "0"
+         shouldTrackSuperviewWidth = "NO">
+      </LoggerValueHistoryTimelineItem>
+      <LoggerValueHistoryTimelineItem
+         documentLocation = "file:///Users/marijnschilling/Projects/TinyConstraints/TinyConstraints.playground/Pages/0%20-%20Tiny%20Constraints.xcplaygroundpage#CharacterRangeLen=14&amp;CharacterRangeLoc=304&amp;EndingColumnNumber=19&amp;EndingLineNumber=9&amp;StartingColumnNumber=5&amp;StartingLineNumber=9&amp;Timestamp=552134501.135277"
+         selectedRepresentationIndex = "0"
+         shouldTrackSuperviewWidth = "NO">
+      </LoggerValueHistoryTimelineItem>
+   </TimelineItems>
+</Timeline>
